Fix bug in deleteItem Func, fix bug in inset item func. Add levelorder F
fix bug in deleteitem where it would not delete if the value was the
main/root element and only had 1 subtree. Also fixed a bug where
insert func would not insert a node if the main root was null. Also
created 2 levelorder function one using an iterative approach and one
using recursion